Moves Time Utils
This might not actually be their final place, since some of these are
currently only used in one place. It gets them out of the main bot
though.
